Improve UX by regrouping all project-related settings within the project properties dialog
